+++ my_portfolio_app/src/App.jsx
@@ -1,6 +1,6 @@
 /* eslint-disable unicorn/filename-case */
 import React, { Suspense } from 'react';
-import { Layout } from './app/Layout.js';
+import { Layout } from './app/layout';
 import { RouterProvider } from './app/providers';
 import { Router } from './pages/index';
 import './18n';
