--- my_portfolio_app/src/app/index.js
@@ -1,1 +1,0 @@
-export * from './Layout.js';
