+++ my_portfolio_app/src/app/index.jsx
@@ -1,0 +1,2 @@
+export { default as Layout } from './layout';
+export { default as RouterProvider } from './providers';
