+++ my_portfolio_app/src/app/layout.jsx
@@ -1,6 +1,6 @@
 /* eslint-disable react/prop-types */
 import React from 'react';
-import { Header } from '../widgets/Header/Header';
+import { Header } from '../widgets/header/header';
 export const Layout = ({ children }) => {
   return (
     <>
