+++ my_portfolio_app/src/entities/PhotoGalleryGrid/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Box } from '@mui/material';
 import { Container } from '@mui/material';
-import style from './photoGalleryGrid.module.css';
+import style from './photoGalleryGrid.module.scss';
 
 export const PhotoGalleryGrid = () => {
   return (
