+++ my_portfolio_app/src/entities/PhotoGalleryGrid/photoGalleryGrid.module.scss
