+++ my_portfolio_app/src/entities/PhotoGalleryHeader/photoGalleryHeader.module.scss
