+++ my_portfolio_app/src/features/UploadPhotoInput/uploadPhotoInput.module.scss
