+++ my_portfolio_app/src/pages/PhotoGallery/photoGallery.jsx
@@ -1,8 +1,8 @@
 import React from 'react';
 import { Box, Container } from '@mui/material';
-import { UploadPhotoInput } from '../../features/UploadPhotoInput';
+import { UploadPhotoInput } from '../../features/uploadPhotoInput';
 import { PhotoGalleryHeader } from '../../entities/PhotoGalleryHeader';
-import { PhotoGalleryGrid } from '../../entities/PhotoGalleryGrid';
+import { PhotoGalleryGrid } from '../../entities/photoGalleryGrid';
 import styles from './photoGallery.module.scss';
 
 const PhotoGallery = () => {
