+++ my_portfolio_app/src/pages/WidgetComponent/widget.jsx
