+++ my_portfolio_app/src/shared/api/weather.jsx
