+++ my_portfolio_app/src/shared/ui/inputError/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Box } from '@mui/material';
 import Lottie from 'lottie-react';
-import AnimationError from '../../../shared/ui/animation/animationError.json';
+import AnimationError from '../animation/animationError.json';
 
 export const InputError = (message) => {
   return (
