+++ my_portfolio_app/src/widgets/ui/animation/animation.jsx
